# docs/swagger.yaml
basePath: /
definitions:
  schema.AuthSwag:
    properties:
      password:
        description: 登录密码
        example: "1234"
        type: string
      user_name:
        description: 登录邮箱
        example: teddy
        type: string
    type: object
...
  schema.CurrentUserSwag:
    properties:
      address:
        example: 广州市白云区xx路2号
        type: string
      email:
        description: 用户名
        example: 123456@qq.com
        type: string
      phone:
        description: 电话
        example: "18677337725"
        type: string
    type: object
  schema.FarmRecordSwag:
    properties:
      end_time:
        example: "1594382165"
        type: string
      info:
        example: 撒了点肥料
        type: string
      oper_name:
        example: 施肥
        type: string
      oper_type:
        example: 施肥
        type: string
      start_time:
        example: "1594382265"
        type: string
    type: object
...
  schema.PasswordSwag:
    properties:
      new_password:
        description: 新密码
        example: "123456"
        type: string
      old_password:
        description: 旧密码
        example: "1234"
        type: string
    type: object
...
  schema.RegSwag:
    properties:
      identity:
        description: identity
        example: "4522261111111111"
        type: string
      password:
        description: 密码
        example: "1234"
        type: string
      username:
        description: 用户名
        example: teddy
        type: string
    required:
    - identity
    - password
    - username
    type: object
  schema.RevokeSwag:
    properties:
      identity:
        example: "4522261111111111"
        type: string
      user_name:
        example: teddy
        type: string
    type: object
...
host: 202.193.60.112:8000
info:
  contact: {}
  description: trace
  license: {}
  title: fabirc bcs API
  version: "1.0"
paths:
...
  /api/v1/user/current:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取登录用户信息
      tags:
      - 用户管理
  /api/v1/user/getHeader:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户头像获取
      tags:
      - 用户管理
  /api/v1/user/getRecords:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户记录数查询
      tags:
      - 用户管理
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名、密码
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/schema.AuthSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 用户登录
      tags:
      - 用户管理
  /api/v1/user/logout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户登出
      tags:
      - 用户管理
  /api/v1/user/operType:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户查询农事数据类型
      tags:
      - 用户管理
  /api/v1/user/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 旧密码、新密码
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/schema.PasswordSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 修改用户密码
      tags:
      - 用户管理
  /api/v1/user/record:
    post:
      consumes:
      - application/json
      parameters:
      - description: 农事记录上传
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/schema.FarmRecordSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户记录上传
      tags:
      - 用户管理
  /api/v1/user/refresh:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 刷新token
      tags:
      - 用户管理
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名、身份证号、密码
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/schema.RegSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: 注册用户
      tags:
      - 用户管理
  /api/v1/user/revoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注销
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/schema.RevokeSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 用户注销
      tags:
      - 用户管理
  /api/v1/user/setHeader:
    post:
      consumes:
      - multipart/form-data
      description: upload file
      parameters:
      - description: 用户头像
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用户更换头像
      tags:
      - 用户管理
  /api/v1/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 字段更新
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/schema.CurrentUserSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 修改登录用户信息
      tags:
      - 用户管理
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"